+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+[v#.#.#] ([month] [YYYY])
+  - Add clone feature to copy projects without the need of exporting and exporting
+
 v4.3.0 (April 2022)
   - No changes